Improve initialization around adding objects to level
Replace passing 1-8 additional float variables into the addObject/Text methods
with an optional table. This method should be less "hacky" as it will allow as
many values as lua handle in a table and all the option values are labeled at
call time, which makes the call more explicit. Update krig3 to replace these
usages as well.