Fix Lua error thrown while traversing options table
Problem:
While traversing a table, the Lua stack does not like it when numbers are read
as strings--the tostring method has the ability to mutate the value on the stack
and this operation will "throw off" its internals when calling the "next"
method.

Solution:
Check the type of both the key and the value when using tostring/number on
either, so that the stack is properly managed.