Fix Lua error thrown when attempting to access a global which is not defined
Problem:
When Lua tries to load a specified global onto the stack and it doesn't exist,
it pushes a nil value onto the top of the stack--depending on future usage of
the stack, this can be problematic and throw an error.

Solution:
Always pop the top of the stack off when the get global call cannot find the
requested value.