Make SoundFx class tolerant to non existing sounds directory
Problem:
The SoundFx class relied on a sounds directory with at least one valid sound
file being present, in order to properly instantiate itself (since the work of
finding and loading sounds is done in the constructor). There were not any
checks to ensure that we were allocating arrays of a length > 0 for the buffers
and sources.

Solution:
By default, set buffers and sources to NULL. If sounds directory is not present
or if no sounds were found, then do not attempt to allocate sources or buffers.
With the sources and buffers properly set to null, we can do null checks to
protect invalid code references.